+++ src/custom-snippets.ts
@@ -87,16 +87,15 @@ async function getSnippetDefinition(codeOrKey: string, editorNS: string, editorR
   const workspaceSnippets = getConfig().customREPLCommandSnippetsWorkspace;
   const workspaceFolderSnippets = getConfig().customREPLCommandSnippetsWorkspaceFolder;
   let snippets = [
-    ...(globalSnippets ? globalSnippets : []),
-    ...(workspaceSnippets ? workspaceSnippets : []),
     ...(workspaceFolderSnippets ? workspaceFolderSnippets : []),
+    ...(workspaceSnippets ? workspaceSnippets : []),
+    ...(globalSnippets ? globalSnippets : []),
   ];
   if (snippets.length < 1) {
     snippets = getConfig().customREPLCommandSnippets;
   }
   const snippetsDict = {};
-  const snippetsKeyDict = {};
-  const snippetsMenuItems: string[] = [];
+  const snippetsMenuItems: vscode.QuickPickItem[] = [];
   snippets.forEach((c: CustomREPLCommandSnippet) => {
     const undefs = ['name', 'snippet'].filter((k) => {
       return !c[k];
@@ -107,11 +106,16 @@ async function getSnippetDefinition(codeOrKey: string, editorNS: string, editorR
     const entry = { ...c };
     entry.ns = entry.ns ? entry.ns : editorNS;
     entry.repl = entry.repl ? entry.repl : editorRepl;
-    const prefix = entry.key !== undefined ? `${entry.key}: ` : '';
-    const item = `${prefix}${entry.name} (${entry.repl})`;
+    const item = {
+      label: `${entry.key ? entry.key + ': ' : ''}${entry.name}`,
+      detail: `${entry.snippet}`,
+      description: `${entry.repl}`,
+      snippet: entry.snippet,
+    };
     snippetsMenuItems.push(item);
-    snippetsDict[item] = entry;
-    snippetsKeyDict[entry.key] = item;
+    if (!snippetsDict[entry.key]) {
+      snippetsDict[entry.key] = entry;
+    }
   });
 
   if (configErrors.length > 0) {
@@ -123,20 +127,20 @@ async function getSnippetDefinition(codeOrKey: string, editorNS: string, editorR
     return;
   }
 
-  let pick: string;
+  let pick: any;
   if (codeOrKey === undefined) {
     // Called without args, show snippets menu
     if (snippetsMenuItems.length > 0) {
       try {
         const pickResult = await util.quickPickSingle({
-          values: snippetsMenuItems.map((a) => ({ label: a })),
+          values: snippetsMenuItems,
           placeHolder: 'Choose a command to run at the REPL',
           saveAs: 'runCustomREPLCommand',
         });
         if (pickResult === undefined || pickResult.label.length < 1) {
           return;
         }
-        pick = pickResult.label; // Assign the label property to pick
+        pick = pickResult;
       } catch (e) {
         console.error(e);
       }
@@ -151,10 +155,10 @@ async function getSnippetDefinition(codeOrKey: string, editorNS: string, editorR
 
   if (pick === undefined) {
     // still no pick, but codeOrKey might be one
-    pick = snippetsKeyDict[codeOrKey];
+    pick = snippetsDict[codeOrKey];
   }
 
-  return pick !== undefined ? snippetsDict[pick] : { snippet: codeOrKey };
+  return pick ?? { snippet: codeOrKey };
 }
 
 export function makeContext(editor: vscode.TextEditor, ns: string, editorNS: string, repl: string) {